Rework files for zc_plugins installation
