+++ Plugin WIP TODO.txt
@@ -1,0 +1,14 @@
+TODO
+Sort out parsing of possible tool locations.
+
+admin\includes\extra_datafiles\backup_mysql.php
+review this
+// Set this to true if the zip options aren't appearing while doing a backup, and you are certain that gzip support exists on your server
+define('COMPRESS_OVERRIDE',false);
+//define('COMPRESS_OVERRIDE',true);
+
+
+Changelog
+Moved tool locations defines to extra_datafiles.
+Converted language file to 158 format, ordered constants
+Moved files to plugins structure
