--- YOUR-ADMIN-FOLDER/includes/extra_datafiles/backup_mysql.php
@@ -1,29 +1,0 @@
-<?php
-//
-// +----------------------------------------------------------------------+
-// |zen-cart Open Source E-commerce                                       |
-// +----------------------------------------------------------------------+
-// | Copyright (c) 2003-2010 The zen-cart developers                           |
-// |                                                                      |
-// | http://www.zen-cart.com/index.php                                    |
-// |                                                                      |
-// | Portions Copyright (c) 2003 osCommerce                               |
-// +----------------------------------------------------------------------+
-// | This source file is subject to version 2.0 of the GPL license,       |
-// | that is bundled with this package in the file LICENSE, and is        |
-// | available through the world-wide-web at the following url:           |
-// | http://www.zen-cart.com/license/2_0.txt.                             |
-// | If you did not receive a copy of the zen-cart license and are unable |
-// | to obtain it through the world-wide-web, please send a note to       |
-// | license@zen-cart.com so we can mail you a copy immediately.          |
-// +----------------------------------------------------------------------+
-//  $Id: backup_mysql.php 154 2010-06-09 13:57:26Z drbyte $
-//
-
-define('FILENAME_BACKUP_MYSQL', 'backup_mysql');
-define('BOX_TOOLS_BACKUP_MYSQL', 'Database Backup - MySQL');
-
-
-// Set this to true if the zip options aren't appearing while doing a backup, and you are certain that gzip support exists on your server
-define('COMPRESS_OVERRIDE',false);
-//define('COMPRESS_OVERRIDE',true);
