+++ files/zc_plugins/BackupMySQL/v2.0.0/admin/includes/languages/english/extra_definitions/lang.backup_mysql_name.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * part of Backup MySQL plugin
+ * @copyright Copyright 2024 Zen Cart Development Team
+ * @license http://www.zen-cart.com/license/2_0.txt GNU Public License V2.0
+ * @version $Id: torvista 2024 Aug 07 $
+ */
+
+$define = [
+     'BOX_TOOLS_BACKUP_MYSQL' => 'MySQL Database Backup-Restore',
+];
+
+return $define;
