+++ files/zc_plugins/BackupMySQL/v2.0.0/manifest.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * part of Backup MySQL plugin
+ * @copyright Copyright 2024 Zen Cart Development Team
+ * @license http://www.zen-cart.com/license/2_0.txt GNU Public License V2.0
+ * @version $Id: torvista 2024 Aug 07 $
+ */
+
+return [
+    'pluginVersion' => 'v2.0.0',
+    'pluginName' => 'Backup MySQL',
+    'pluginDescription' => 'Backup and Restore the MySQL database',
+    'pluginAuthor' => 'Dr.Byte and others',
+    'pluginId' => 7, // ID from Zen Cart forum
+    'zcVersions' => ['v158'],
+    'changelog' => 'https://github.com/torvista/Zen_Cart-Backup_MYSQL_Plugin/tags', // online URL (eg github release tag page, or changelog file there) or local filename only, ie: changelog.txt (in same dir as this manifest file)
+    'github_repo' => 'https://github.com/torvista/Zen_Cart-Backup_MYSQL_Plugin', // url
+    'pluginGroups' => [],
+];
